chore: move files out of src folder
all but one file (request-key) was being accessed externally, no need for the indirection
