+++ framed.js
@@ -1,3 +1,23 @@
 'use strict';
 
-module.exports = require('./src/framed');
+var Client = require('ifrau/client'),
+	Promise = require('lie');
+
+var REQUEST_KEY = require('./request-key');
+
+module.exports = function getFramedJwt(scope) {
+	return Promise
+		.resolve()
+		.then(function() {
+			var client = new Client({
+				resizeFrame: false,
+				syncTitle: false
+			});
+
+			return client
+				.connect()
+				.then(function() {
+					return client.request(REQUEST_KEY, scope);
+				});
+		});
+};
