+++ host.js
@@ -1,3 +1,9 @@
 'use strict';
 
-module.exports = require('./src/host');
+var getJwt = require('./local');
+
+var REQUEST_KEY = require('./request-key');
+
+module.exports = function ifrauJwtHost(host) {
+	host.onRequest(REQUEST_KEY, getJwt);
+};
