--- src/host.js
@@ -1,9 +1,0 @@
-'use strict';
-
-var getJwt = require('./local');
-
-var REQUEST_KEY = require('./request-key');
-
-module.exports = function ifrauJwtHost(host) {
-	host.onRequest(REQUEST_KEY, getJwt);
-};
