+++ test/framed.spec.js
@@ -1,3 +1,3 @@
 'use strict';
 
-require('../src/framed');
+require('../framed');
